r9035/sofa-dev : FIX: parsing of filenames with OglShader in 64bits windows. long types remain 32bit wide in windows x64 so conversion from size_t to unsigned long is not valid. I presume there are other traps like that kicking around...

+++ modules/sofa/component/visualmodel/OglShader.cpp
@@ -91,8 +91,8 @@ void OglShader::init()
     std::string tempStr = vertFilename.getFullPath();
     std::string file;
     const std::string SEPARATOR = ";";
-    unsigned long int pos = 0;
-    unsigned long int oldPos = 0;
+    size_t pos = 0;
+    size_t oldPos = 0;
 
     pos = tempStr.find(SEPARATOR, oldPos);
